chore: move presenter viewmodels into separate files
