+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedLoadingViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Nikhil Menon on 2/21/24.
+//
+
+import Foundation
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
