+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedPresenter.swift
@@ -7,16 +7,6 @@
 
 import Foundation
 
-public struct FeedErrorViewModel {
-    public let message: String?
-    
-    static var noError = FeedErrorViewModel(message: nil)
-    
-    static func error(message: String) -> FeedErrorViewModel {
-        return FeedErrorViewModel(message: message)
-    }
-}
-
 public protocol FeedErrorView {
     func display(_ viewModel: FeedErrorViewModel)
 }
@@ -27,14 +17,6 @@ public protocol FeedLoadingView {
 
 public protocol FeedView {
     func display(_ viewModel: FeedViewModel)
-}
-
-public struct FeedViewModel {
-    public let feed: [FeedImage]
-}
-
-public struct FeedLoadingViewModel {
-    public let isLoading: Bool
 }
 
 public final class FeedPresenter {
