+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedViewModel.swift
+//  EssentialFeed
+//
+//  Created by Nikhil Menon on 2/21/24.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
